requirements.txt
fastapi
redis
types-redis
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
email-validator
bcrypt
paho-mqtt